requirements.txt
# Core Data Analysis Stack
pandas==2.1.3
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Streamlit App
streamlit==1.29.0
streamlit-extras==0.7.1
python-dotenv==1.0.0
jinja2==3.1.2
watchdog==6.0.0

# Database and Scheduling
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymysql==1.1.0
apscheduler==3.10.4

# Google API
gspread==5.12.3
oauth2client==4.1.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.172.0

# ML & Whisper
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
torch==2.7.1
soundfile==0.13.1
pyttsx3==2.98

# Others
requests==2.31.0
protobuf==4.25.8
setuptools==80.9.0
wheel==0.45.1
Pillow==10.1.0


